spike(PortDemo): Added easier to explain interaction test
