+++ src/lib/AnalysisApps/GWAS/Components/SelectCohort/SelectCohort.stories.tsx
@@ -69,28 +69,6 @@ export const MockedSuccess: Story = {
     },
   },
   render: () => <SelectCohortWithHooks />, // see https://storybook.js.org/docs/writing-stories
-  /* Interaction Tests */
-  play: async ({ canvasElement }) => {
-    const canvas = within(canvasElement);
-
-    // 👇 Simulate interactions with the component
-    // Simple check to make sure Search by cohort name filters
-    await expect(canvas.getByTestId('GWASUI-mainTable')).toBeInTheDocument();
-
-    await userEvent.click(canvas.getAllByRole('textbox')[0]);
-
-    await userEvent.type(
-      canvas.getAllByRole('textbox')[0],
-      'team2 - test new cohort - catch all',
-    );
-
-    await expect(
-      canvas.getByText('team2 - test new cohort - catch all'),
-    ).toBeInTheDocument();
-    await expect(
-      canvas.queryByText('team2 - test new cohort - medium + large'),
-    ).not.toBeInTheDocument();
-  },
 };
 
 export const MockedError: Story = {
